+++ ceammc/ext/src/ui/ui_gain.h
@@ -56,6 +56,8 @@ public:
     void onMouseDrag(t_object* view, const t_pt& pt, long modifiers);
     void onDblClick(t_object* view, const t_pt& pt, long modifiers);
     void onMouseWheel(const t_pt& pt, long modifiers, float delta);
+    void onPopup(t_symbol* menu_name, long item_idx);
+    void showPopup(const t_pt& pt, const t_pt& abs_pt);
 
     t_float dbValue() const;
     t_float ampValue() const;