+++ ceammc/CHANGELOG.md
@@ -23,6 +23,7 @@
 - ui.gain~ and ui.gain2~ @relative property added for relative input mode (by default)
 - ui.gain~ and ui.gain2~ slow value change added in relative mode while holding SHIFT
 - ui.gain~/ui.gain2~ mouse wheel support added (with slow on SHIFT)
+- ui.gain~/ui.gain2~ popup added
 - ui.polar added
 - ui.slider slow change with SHIFT added
 - ui.slider log scale support added with @scale property